requirements.txt
appnope==0.1.0
aubio==0.4.6
beautifulsoup4==4.6.0
bleach==1.5.0
brython==3.4.0
bs4==0.0.1
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
cycler==0.10.0
decorator==4.1.2
Django==1.11.7
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
docopt==0.6.2
easygui==0.98.1
entrypoints==0.2.3
enum-compat==0.0.2
enum34==1.1.6
eventlet==0.21.0
Flask==0.12.2
Flask-Session==0.3.1
Flask-SocketIO==2.9.3
fuckit==4.8.0
greenlet==0.4.12
grip==4.3.2
html5lib==0.9999999
idna==2.6
imutils==0.4.3
ipykernel==4.6.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.5
itsdangerous==0.24
jedi==0.11.0
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.0.9
Markdown==2.6.9
MarkupSafe==1.0
matplotlib==2.1.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.2
numpy==1.14.0
olefile==0.44
opencv-python==3.3.0.10
pandas==0.21.0
pandas-datareader==0.5.0
pandocfilters==1.4.2
parso==0.1.0
path-and-address==2.0.1
pexpect==4.3.0
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
protobuf==3.4.0
psycopg2==2.7.3.2
ptyprocess==0.5.2
PyAudio==0.2.11
pygame==1.9.3
Pygments==2.2.0
PyJWT==1.5.3
pyparsing==2.2.0
python-dateutil==2.6.1
python-engineio==2.0.2
python-socketio==1.8.4
pytz==2017.2
PyYAML==3.12
pyzmq==16.0.3
qtconsole==4.3.1
redis==2.10.6
requests==2.18.4
requests-file==1.4.2
requests-ftp==0.3.1
scipy==1.0.0
selenium==3.5.0
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.1
tensorflow==1.4.0
tensorflow-tensorboard==0.4.0rc2
terminado==0.8.1
testpath==0.3.1
tornado==4.5.2
tqdm==4.19.4
traitlets==4.3.2
urllib3==1.22
virtualenv==15.1.0
wcwidth==0.1.7
Werkzeug==0.12.2
widgetsnbextension==3.0.8
